backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.10.4
attrs==24.3.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
dnspython==2.6.1
Flask==3.0.3
Flask-APScheduler==1.13.1
flask-caching==2.1.5
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
flask-cors==4.0.2
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.9
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
realtime==2.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
yarl==1.18.3
